finished

+++ admin/sql/bajar_mysql.php
@@ -17,37 +17,56 @@
             if($resultado_destiny->num_rows != 0) {
             
 
-                $sql = "SELECT cuentas.fecha_hora 
-                        FROM transacciones INNER JOIN cuentas ON transacciones.id_cuenta_origen = cuentas.id_usuario
-                        WHERE timestampdiff(month, cuentas.fecha_hora, NOW()) > 3 
-                        AND alias = '" . $_POST['destiny'] . "';";
+                $sql = "SELECT id FROM cuentas WHERE alias = '" . $_POST['destiny'] . "'";
+
+                $resultado = mysqli_query($conexion, $sql);
+                $resultado_id = mysqli_fetch_array($resultado)[0];
+
+                $sql = "SELECT * 
+                FROM transacciones 
+                WHERE id_cuenta_destino = $resultado_id || id_cuenta_origen = $resultado_id";
 
                 $resultado = mysqli_query($conexion, $sql);
 
-                if($resultado->num_rows != 0) {
+                //Valido si la cuenta hallada no tiene transacciones
+                if($resultado->num_rows == 0) {
 
-                    mysqli_query($conexion, "SET foreign_key_checks = 0");
-    
-                    $resultado = mysqli_query($conexion, "DELETE from cuentas 
-                    where alias='" . $_POST['destiny'] . "'");
-    
-                    mysqli_query($conexion, "SET foreign_key_checks = 1");
-                    if($resultado) {
-    
-                    $_SESSION['sucess'] = "Baja realizada";
-                    $conexion -> commit();
-                    header("Location: ../admin.php");
-                    die();
-    
-                    } else {
-                        $conexion -> rollback();
-                        $_SESSION['error'] = "Error al dar de baja";
-                        header("Location: ../admin.php");
-                    }
+
+                    $sql = "SELECT fecha_hora 
+                        FROM cuentas
+                        WHERE timestampdiff(month, fecha_hora, NOW()) >= 0 
+                        AND alias = '" . $_POST['destiny'] . "';";
+
+                    $resultado = mysqli_query($conexion, $sql);
+
+                    //Valido si el cliente tiene más de 3 meses de antiguedad
+                        if($resultado->num_rows != 0) {
+
+                            mysqli_query($conexion, "SET foreign_key_checks = 0");
+            
+                            $resultado = mysqli_query($conexion, "DELETE from cuentas 
+                            where alias='" . $_POST['destiny'] . "'");
+            
+                            mysqli_query($conexion, "SET foreign_key_checks = 1");
+                            if($resultado) {
+            
+                            $_SESSION['sucess'] = "Baja realizada";
+                            $conexion -> commit();
+                            header("Location: ../admin.php");
+                            die();
+        
+                        } else {
+                            $conexion -> rollback();
+                            $_SESSION['error'] = "Error al dar de baja";
+                            header("Location: ../admin.php");
+                        }
                             
+                    } else {
+                        $_SESSION['error'] = "La cuenta tiene menos de 3 meses de antiguedad";
+                            
+                    }   
                 } else {
-                    $_SESSION['error'] = "La última transferencia de la cuenta con este alias fue hace menos de 3 meses";
-                        
+                    $_SESSION['error'] = "La cuenta a eliminar cuenta con transferencias o depositos hechos";
                 }
             
             } else {
